Extract FilterExpressionASTFactory
